Make iterating over entities more efficient.
EntityPool now manages its own array instead of using ArrayPool.
In benchmark mode, we were spending 27.6% of runtime in the
next() and hasNext() methods of EntityPoolIterator due to large
numbers of calls to ArrayPool.get(). With this change, the total
execution time of those two EntityPooliterator methods is
roughly halved. (They're now 18.7% of total runtime.)